handoff/20250928/40_App/morningai_enhanced/api-backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
PyJWT==2.8.0
cryptography==41.0.4
packaging==23.1
PyYAML==6.0.1
requests==2.31.0
gunicorn==21.2.0
redis==4.6.0
python-dotenv==1.0.0
openai==0.28.1

